requirements.txt
configparser
mysql-connector
pyyaml
requests
pprint
